Reworked input system

+++ Source/EngineMain/Core/Engine/GameEngine.h
@@ -16,7 +16,7 @@ struct EngineTime
     /* Global time dilation */
     float timeDilation = 1;
     /* Time dilation for activity of app 1 if app is active 0 if app is in background not active*/
-    int32 activeTimeDilation = 1;
+    float activeTimeDilation = 1;
 
     void engineStart();
     void tickStart();
